refactor: 사용자 인증/계정 패키지 리팩토링
